frontend/backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
langchain
langchain-google-genai
python-multipart
supabase
